external_repos/dhd-pluginlake/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plugindb"
version = "0.1.0"
description = "Local datalake for plugin projects"
readme = "README.md"
requires-python = "==3.12.4"
authors = [
    { name = "Yannick Vinkesteijn", email = "y.vinkesteijn@dhd.nl" },
]
dependencies = [
    "pyyaml",
    "fastapi",
    "polars",
    "uvicorn",
    "aiofiles",
    "pyarrow",
    "python-dotenv",
    "python-multipart"
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "dask-expr",
    "pytest-pythonpath",
    "nbstripout",
    "ruff"
]

[tool.pytest.ini_options]
pythonpaths = ["src", "pluginlake"]
addopts = [
    "--import-mode=importlib",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
lint.select = ["E", "F", "W", "C90", "ANN"]
lint.ignore = ["E203", "E501", "ANN101", "ANN204", "ANN002", "ANN003"]
exclude = [
    ".git",
    "__pycache__",
    ".ipynb",
    "__init__.py",
    "tests/*",
    "*.ipynb"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["plugindb"]
force-wrap-aliases = true
combine-as-imports = true
split-on-trailing-comma = true
order-by-type = true
case-sensitive = false